backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow
opencv-python-headless
pillow
numpy
python-multipart
fastapi-cors 